Move tests to framework project
